Rename `utils` submodule `time` -> `timeutils`.
This prevents clashes with the python standard library `time`.
